Moved new serialization code to 'Current/Serialization'.
