+++ Reflection/test.cc
@@ -29,11 +29,6 @@ SOFTWARE.
 #include "reflection.h"
 #include "../Bricks/strings/join.h"
 
-// TODO(dkorolev): Use RapidJSON from outside Cereal.
-#include "../3rdparty/cereal/include/external/rapidjson/document.h"
-#include "../3rdparty/cereal/include/external/rapidjson/prettywriter.h"
-#include "../3rdparty/cereal/include/external/rapidjson/genericstream.h"
-
 #include "../3rdparty/gtest/gtest-main.h"
 
 namespace reflection_test {
@@ -185,19 +180,9 @@ TEST(Reflection, CurrentStructInternals) {
   EXPECT_EQ(1u, FieldCounter<DerivedFromFoo>::value);
 }
 
+// TODO: deprecate this test?
+#if 0
 namespace reflection_test {
-
-CURRENT_STRUCT(Serializable) {
-  CURRENT_FIELD(i, uint64_t);
-  CURRENT_FIELD(s, std::string);
-};
-
-CURRENT_STRUCT(ComplexSerializable) {
-  CURRENT_FIELD(j, uint64_t);
-  CURRENT_FIELD(q, std::string);
-  CURRENT_FIELD(v, std::vector<std::string>);
-  CURRENT_FIELD(z, Serializable);
-};
 
 struct CollectFieldNames {
   std::vector<std::string>& output_;
@@ -234,247 +219,4 @@ TEST(Reflection, VisitAllFields) {
   }
 }
 
-namespace reflection_json {
-
-template <typename T>
-struct SaveIntoJSONImpl;
-
-template <typename T>
-struct AssignToRapidJSONValueImpl {
-  static void WithDedicatedStringTreatment(rapidjson::Value& destination, const T& value) {
-    destination = value;
-  }
-};
-
-template <>
-struct AssignToRapidJSONValueImpl<std::string> {
-  static void WithDedicatedStringTreatment(rapidjson::Value& destination, const std::string& value) {
-    destination.SetString(value.c_str(), value.length());
-  }
-};
-
-template <typename T>
-void AssignToRapidJSONValue(rapidjson::Value& destination, const T& value) {
-  AssignToRapidJSONValueImpl<T>::WithDedicatedStringTreatment(destination, value);
-}
-
-#define CURRENT_DECLARE_PRIMITIVE_TYPE(unused_typeid_index, cpp_type, unused_current_type) \
-  template <>                                                                              \
-  struct SaveIntoJSONImpl<cpp_type> {                                                      \
-    static void Go(rapidjson::Value& destination,                                          \
-                   rapidjson::Document::AllocatorType&,                                    \
-                   const cpp_type& value) {                                                \
-      AssignToRapidJSONValue(destination, value);                                          \
-    }                                                                                      \
-  };
-#include "primitive_types.dsl.h"
-#undef CURRENT_DECLARE_PRIMITIVE_TYPE
-
-template <typename T>
-struct SaveIntoJSONImpl<std::vector<T>> {
-  static void Go(rapidjson::Value& destination,
-                 rapidjson::Document::AllocatorType& allocator,
-                 const std::vector<T>& value) {
-    destination.SetArray();
-    rapidjson::Value element_to_push;
-    for (const auto& element : value) {
-      SaveIntoJSONImpl<T>::Go(element_to_push, allocator, element);
-      destination.PushBack(element_to_push, allocator);
-    }
-  }
-};
-
-// TODO(dkorolev): A smart `enable_if` to not treat any non-primitive type as a `CURRENT_STRUCT`?
-template <typename T>
-struct SaveIntoJSONImpl {
-  struct Visitor {
-    rapidjson::Value& destination_;
-    rapidjson::Document::AllocatorType& allocator_;
-
-    Visitor(rapidjson::Value& destination, rapidjson::Document::AllocatorType& allocator)
-        : destination_(destination), allocator_(allocator) {}
-
-    // IMPORTANT: Pass in `const char* name`, as `const std::string& name`
-    // would fail memory-allocation-wise due to over-smartness of RapidJSON.
-    template <typename U>
-    void operator()(const char* name, const U& value) const {
-      rapidjson::Value placeholder;
-      SaveIntoJSONImpl<U>::Go(placeholder, allocator_, value);
-      destination_.AddMember(name, placeholder, allocator_);
-    }
-  };
-
-  static void Go(rapidjson::Value& destination, rapidjson::Document::AllocatorType& allocator, const T& value) {
-    destination.SetObject();
-
-    Visitor serializer(destination, allocator);
-    current::reflection::VisitAllFields<bricks::decay<T>, current::reflection::FieldNameAndImmutableValue>::
-        WithObject(value, serializer);
-  }
-};
-
-template <typename T>
-std::string JSON(const T& value) {
-  rapidjson::Document document;
-  rapidjson::Value& destination = document;
-
-  SaveIntoJSONImpl<T>::Go(destination, document.GetAllocator(), value);
-
-  std::ostringstream os;
-  auto stream = rapidjson::GenericWriteStream(os);
-  auto writer = rapidjson::Writer<rapidjson::GenericWriteStream>(stream);
-  document.Accept(writer);
-
-  return os.str();
-}
-
-}  // namespace reflection_json
-
-TEST(Reflection, SerializeIntoJSON) {
-  using namespace reflection_test;
-  using namespace reflection_json;
-
-  // Simple serialization.
-  Serializable simple_object;
-  simple_object.i = 0;
-  simple_object.s = "";
-
-  EXPECT_EQ("{\"i\":0,\"s\":\"\"}", JSON(simple_object));
-
-  simple_object.i = 42;
-  simple_object.s = "foo";
-  EXPECT_EQ("{\"i\":42,\"s\":\"foo\"}", JSON(simple_object));
-
-  // Nested serialization.
-  ComplexSerializable complex_object;
-  complex_object.j = 43;
-  complex_object.q = "bar";
-  complex_object.v.push_back("one");
-  complex_object.v.push_back("two");
-  complex_object.z = simple_object;
-
-  EXPECT_EQ("{\"j\":43,\"q\":\"bar\",\"v\":[\"one\",\"two\"],\"z\":{\"i\":42,\"s\":\"foo\"}}",
-            JSON(complex_object));
-
-  // Complex serialization makes a copy.
-  simple_object.i = -1000;
-  EXPECT_EQ(42ull, complex_object.z.i);
-  EXPECT_EQ("{\"j\":43,\"q\":\"bar\",\"v\":[\"one\",\"two\"],\"z\":{\"i\":42,\"s\":\"foo\"}}",
-            JSON(complex_object));
-}
-
-TEST(Reflection, LOLWUT) {
-  // RapidJSON requires the top-level node to be an array or object.
-  // Thus, just `JSON(42)` or `JSON("foo")` doesn't work. But arrays do. ;-)
-  using namespace reflection_json;
-  EXPECT_EQ("[1,2,3]", JSON(std::vector<uint64_t>({1, 2, 3})));
-  EXPECT_EQ("[[\"one\",\"two\"],[\"three\",\"four\"]]",
-            JSON(std::vector<std::vector<std::string>>({{"one", "two"}, {"three", "four"}})));
-}
-
-// RapidJSON examples framed as tests. One day we may wish to remove them. -- D.K.
-TEST(RapidJSON, Smoke) {
-  using rapidjson::Document;
-  using rapidjson::Value;
-  using rapidjson::Writer;
-  using rapidjson::GenericWriteStream;
-
-  std::string json;
-
-  {
-    Document document;
-    auto& allocator = document.GetAllocator();
-    Value foo("bar");
-    document.SetObject().AddMember("foo", foo, allocator);
-
-    EXPECT_TRUE(document.IsObject());
-    EXPECT_FALSE(document.IsArray());
-    EXPECT_TRUE(document.HasMember("foo"));
-    EXPECT_TRUE(document["foo"].IsString());
-    EXPECT_EQ("bar", document["foo"].GetString());
-
-    std::ostringstream os;
-    auto stream = GenericWriteStream(os);
-    auto writer = Writer<GenericWriteStream>(stream);
-    document.Accept(writer);
-    json = os.str();
-  }
-
-  EXPECT_EQ("{\"foo\":\"bar\"}", json);
-
-  {
-    Document document;
-    ASSERT_FALSE(document.Parse<0>(json.c_str()).HasParseError());
-    EXPECT_TRUE(document.IsObject());
-    EXPECT_TRUE(document.HasMember("foo"));
-    EXPECT_TRUE(document["foo"].IsString());
-    EXPECT_EQ(std::string("bar"), document["foo"].GetString());
-    EXPECT_FALSE(document.HasMember("bar"));
-    EXPECT_FALSE(document.HasMember("meh"));
-  }
-}
-
-TEST(RapidJSON, Array) {
-  using rapidjson::Document;
-  using rapidjson::Value;
-  using rapidjson::Writer;
-  using rapidjson::GenericWriteStream;
-
-  std::string json;
-
-  {
-    Document document;
-    auto& allocator = document.GetAllocator();
-    document.SetArray();
-    Value element;
-    element = 42;
-    document.PushBack(element, allocator);
-    element = "bar";
-    document.PushBack(element, allocator);
-
-    EXPECT_TRUE(document.IsArray());
-    EXPECT_FALSE(document.IsObject());
-
-    std::ostringstream os;
-    auto stream = GenericWriteStream(os);
-    auto writer = Writer<GenericWriteStream>(stream);
-    document.Accept(writer);
-    json = os.str();
-  }
-
-  EXPECT_EQ("[42,\"bar\"]", json);
-}
-
-TEST(RapidJSON, NullInString) {
-  using rapidjson::Document;
-  using rapidjson::Value;
-  using rapidjson::Writer;
-  using rapidjson::GenericWriteStream;
-
-  std::string json;
-
-  {
-    Document document;
-    auto& allocator = document.GetAllocator();
-    Value s;
-    s.SetString("terrible\0avoided", strlen("terrible") + 1 + strlen("avoided"));
-    document.SetObject();
-    document.AddMember("s", s, allocator);
-
-    std::ostringstream os;
-    auto stream = GenericWriteStream(os);
-    auto writer = Writer<GenericWriteStream>(stream);
-    document.Accept(writer);
-    json = os.str();
-  }
-
-  EXPECT_EQ("{\"s\":\"terrible\\u0000avoided\"}", json);
-
-  {
-    Document document;
-    ASSERT_FALSE(document.Parse<0>(json.c_str()).HasParseError());
-    EXPECT_EQ(std::string("terrible"), document["s"].GetString());
-    EXPECT_EQ(std::string("terrible\0avoided", strlen("terrible") + 1 + strlen("avoided")),
-              std::string(document["s"].GetString(), document["s"].GetStringLength()));
-  }
-}
+#endif
